refactor: wallet-store module
