+++ packages/client-utils/test/signing-smart-wallet-kit.test.ts
@@ -8,10 +8,7 @@ import type {
   SigningStargateClient,
   StdFee,
 } from '@cosmjs/stargate';
-import {
-  makeSigningSmartWalletKit,
-  reflectWalletStore,
-} from '../dist/signing-smart-wallet-kit.js';
+import { makeSigningSmartWalletKit } from '../dist/signing-smart-wallet-kit.js';
 import { LOCAL_CONFIG } from '../src/network-config.js';
 import { makeSmartWalletKit } from '../src/smart-wallet-kit.js';
 
@@ -224,52 +221,4 @@ test('sendBridgeAction with signerData supports custom fee', async t => {
     signerData,
   });
   t.is(broadcastCalls.length, 1);
-});
-
-test('reflectWalletStore save/overwrite chains without changing args', async t => {
-  const actions: any[] = [];
-  let lastId = '';
-  const signer = {
-    sendBridgeAction: async action => {
-      lastId = action.message?.id || '';
-      actions.push(action);
-      return { code: 0 };
-    },
-    query: {
-      getLastUpdate: async () => ({
-        updated: 'invocation',
-        id: lastId,
-        result: 'ok',
-      }),
-    },
-  };
-
-  const walletStore = reflectWalletStore(
-    // @ts-expect-error minimal signer for test
-    signer,
-    { setTimeout: globalThis.setTimeout, makeNonce: () => '123' },
-  );
-
-  const target = walletStore.get<any>('ymaxControl');
-
-  await target.getCreatorFacet('arg1', 2);
-  t.is(actions.length, 1);
-  t.deepEqual(actions[0].message.args, ['arg1', 2]);
-  t.falsy(actions[0].message.saveResult);
-
-  await target.saveAs('creatorFacet').getCreatorFacet('arg2');
-  t.is(actions.length, 2);
-  t.deepEqual(actions[1].message.args, ['arg2']);
-  t.deepEqual(actions[1].message.saveResult, {
-    name: 'creatorFacet',
-    overwrite: false,
-  });
-
-  await target.overwrite('creatorFacet').getCreatorFacet('arg3');
-  t.is(actions.length, 3);
-  t.deepEqual(actions[2].message.args, ['arg3']);
-  t.deepEqual(actions[2].message.saveResult, {
-    name: 'creatorFacet',
-    overwrite: true,
-  });
 });
